refactor(parser): extract active token clone
